fix: renamed gameEnd to GameEvents
