+++ src/styles/index.scss
@@ -6,6 +6,6 @@
   '../Components/Registration/registration.scss',
   '../Components/Login/login.scss', '../Components/Preloader/preloader.scss',
   '../Components/SelectWord/index.SelectWord.scss',
-  '../Components/GameEnd/index.GameEnd.scss',
+  '../Components/Game/GameEvents/index.GameEvents.scss',
   '../Components/PageLayout/styles.PageLayout.scss';
 @import '../SocketIoClient/style.chat.scss';
